feat: introduce OriginHttpClientManager bean for origin configuration
